+++ inlang/source-code/paraglide/paraglide-next/examples/app/src/app/layout.tsx
@@ -1,11 +1,12 @@
 import "@/lib/ui/styles.css"
-import { LanguageProvider } from "@inlang/paraglide-next"
+import { LanguageProvider, generateAlternateLinks } from "@inlang/paraglide-next"
 import { AvailableLanguageTag, languageTag } from "@/paraglide/runtime"
 import { Header } from "@/lib/ui/Header"
 import * as m from "@/paraglide/messages.js"
-import type { Metadata } from "next"
+import type { Metadata, ResolvingMetadata } from "next"
+import { strategy } from "@/lib/i18n"
 
-export function generateMetadata(): Metadata {
+export function generateMetadata(_: unknown, parent: ResolvingMetadata): Metadata {
 	const locale = languageTag()
 	return {
 		title: m.paraglide_and_next_app_router(),
@@ -13,6 +14,13 @@ export function generateMetadata(): Metadata {
 		icons: "/favicon.png",
 		openGraph: {
 			locale,
+		},
+		alternates: {
+			languages: generateAlternateLinks({
+				origin: "https://example.com",
+				strategy: strategy,
+				resolvingMetadata: parent,
+			}),
 		},
 	}
 }