+++ inlang/source-code/paraglide/paraglide-next/src/app/index.client.tsx
@@ -2,7 +2,7 @@
 export { Middleware } from "./middleware"
 export { Navigation } from "./navigation/navigation.client.js"
 export { initializeLanguage } from "./initializeLanguage.client.js"
-export { generateAlternateLinks } from "./utils/alternate-links.js"
+export { generateAlternateLinks } from "./alternateLinks.js"
 
 // Routing Strategies
 export * from "./routing-strategy/strategies"