+++ inlang/source-code/paraglide/paraglide-next/src/app/routing-strategy/strats/prefixStrategy.ts
@@ -8,7 +8,6 @@ import {
 	resolveUserPathDefinitions,
 } from "@inlang/paraglide-js/internal/adapter-utils"
 import type { RoutingStrategy } from "../interface"
-import { usePrefixDetection } from "../../middleware/detection/prefixDetection"
 import { DEV } from "../../env"
 import { rsc } from "rsc-env"
 import { availableLanguageTags, sourceLanguageTag } from "$paraglide/runtime.js"
@@ -23,6 +22,7 @@ export function PrefixStrategy<T extends string>({
 	pathnames: userPathnames,
 	exclude,
 	prefixDefault,
+	prefixes,
 }: {
 	exclude?: (path: string) => boolean
 	pathnames?: UserPathDefinitionTranslations<T>
@@ -40,10 +40,50 @@ export function PrefixStrategy<T extends string>({
 	 *
 	 */
 	prefixDefault?: "always" | "never"
+
+	/**
+	 * Custom language prefixes
+	 * Do not include slashes in the prefixes
+	 *
+	 * @example
+	 * ```ts
+	 * prefixes: {
+	 *   "de-CH": "ch",
+	 *   "en-US": "us",
+	 * }
+	 * ```
+	 */
+	prefixes?: {
+		[key in T]?: string
+	}
 } = {}): RoutingStrategy<T> {
 	const resolvedPathnames = /** @__PURE__ */ userPathnames
 		? resolveUserPathDefinitions(userPathnames, availableLanguageTags)
 		: {}
+
+	const resolvedPrefixes = {
+		...(Object.fromEntries(availableLanguageTags.map((lang) => [lang, lang])) as Record<T, string>),
+		...prefixes,
+	}
+
+	if (DEV) {
+		const withSlashes = Object.values(resolvedPrefixes).filter(
+			(val) => typeof val === "string" && val.includes("/")
+		)
+		if (withSlashes.length) {
+			throw new Error(
+				["[parglide-sveltekit:dev-only] Prefixes on cannot include slashes.", ...withSlashes].join(
+					"\n"
+				)
+			)
+		}
+
+		const paths = Object.values(resolvedPathnames)
+		const hasDuplicates = new Set(paths).size !== paths.length
+		if (hasDuplicates) {
+			throw new Error("[parglide-sveltekit:dev-only] `prefixes` must be unique.")
+		}
+	}
 
 	prefixDefault ??= "never"
 
@@ -60,8 +100,13 @@ export function PrefixStrategy<T extends string>({
 	}
 
 	function getCanonicalPath(localisedPath: `/${string}`, locale: T): `/${string}` {
-		const pathWithoutLocale: `/${string}` = localisedPath.startsWith(`/${locale}`)
-			? ((localisedPath.replace(`/${locale}`, "") || "/") as `/${string}`)
+		// get the prefix for this language
+		const prefix = resolvedPrefixes[locale] ?? locale
+
+		const pathWithoutLocale: `/${string}` = localisedPath.startsWith(`/${prefix}/`)
+			? (localisedPath.replace(`/${prefix}`, "") as `/${string}`)
+			: localisedPath === `/${prefix}`
+			? "/"
 			: localisedPath
 
 		for (const [canonicalPathDefinition, translationsForPath] of Object.entries(
@@ -108,8 +153,10 @@ export function PrefixStrategy<T extends string>({
 			const translatedPath = getTranslatedPath(canonicalPath, targetLanguage, resolvedPathnames)
 			const shouldAddPrefix = targetLanguage !== sourceLanguageTag || prefixDefault === "always"
 
+			const prefix = resolvedPrefixes[targetLanguage] ?? targetLanguage
+
 			const localisedPath: `/${string}` = shouldAddPrefix
-				? `/${targetLanguage}${translatedPath == "/" ? "" : translatedPath}`
+				? `/${prefix}${translatedPath == "/" ? "" : translatedPath}`
 				: translatedPath
 			return {
 				pathname: localisedPath,
@@ -118,8 +165,14 @@ export function PrefixStrategy<T extends string>({
 		getCanonicalPath,
 
 		resolveLocale(request) {
-			const detect = usePrefixDetection({ availableLanguageTags })
-			const detected = detect(request) as T | undefined
+			const pathWithoutBase = request.nextUrl.pathname
+
+			const entries = Object.entries(resolvedPrefixes) as [T, string][]
+			const entry = entries.find(
+				([, prefix]) =>
+					pathWithoutBase.startsWith(`/${prefix}/`) || pathWithoutBase === `/${prefix}`
+			)
+			const detected = entry?.[0]
 
 			// If no prefix is detected and prefixDefault is "never" -> use default language
 			// Otherwise leave it ambiguous