+++ inlang/source-code/paraglide/paraglide-unplugin/src/index.ts
@@ -9,7 +9,7 @@ import {
 import { openRepository, findRepoRoot } from "@lix-js/client"
 import path from "node:path"
 import fs from "node:fs/promises"
-import { compile, writeOutput, Logger } from "@inlang/paraglide-js/internal"
+import { compile, writeOutput, Logger, classifyProjectErrors } from "@inlang/paraglide-js/internal"
 import crypto from "node:crypto"
 
 const PLUGIN_NAME = "unplugin-paraglide"
@@ -46,7 +46,7 @@ export const paraglide = createUnplugin((config: UserConfig) => {
 		if (currentMessagesHash === previousMessagesHash) return
 
 		if (messages.length === 0) {
-			logger.warn(`No messages found - Skipping compilation into ${options.outdir}`)
+			logger.warn("No messages found - Skipping compilation")
 			return
 		}
 
@@ -107,8 +107,26 @@ export const paraglide = createUnplugin((config: UserConfig) => {
 			async buildStart() {
 				const project = await getProject()
 
-				//Always fully compile once on build start
-				await triggerCompile(project.query.messages.getAll(), project.settings())
+				const initialMessages = project.query.messages.getAll()
+				const settings = project.settings()
+				await triggerCompile(initialMessages, settings)
+
+				project.errors.subscribe((errors) => {
+					if (errors.length === 0) return
+
+					const { fatalErrors, nonFatalErrors } = classifyProjectErrors(errors)
+					for (const error of nonFatalErrors) {
+						logger.warn(error.message)
+					}
+
+					for (const error of fatalErrors) {
+						if (error instanceof Error) {
+							logger.error(error.message) // hide the stack trace
+						} else {
+							logger.error(error)
+						}
+					}
+				})
 
 				let numInvocations = 0
 				project.query.messages.getAll.subscribe((messages) => {