+++ inlang/source-code/editor/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @inlang/editor
 
+## 0.1.8
+
+### Patch Changes
+
+- a377572: Update app server settings
+
+## 0.1.7
+
+### Patch Changes
+
+- c07c5a4: fix: Ninja recommendation
+- Updated dependencies [e8d74b9]
+  - @lix-js/fs@2.2.0
+  - @lix-js/client@2.2.1
+  - @inlang/settings-component@1.0.30
+
+## 0.1.6
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.29
+
+## 0.1.5
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.28
+
+## 0.1.4
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.27
+
+## 0.1.3
+
+### Patch Changes
+
+- 9886cfa: show project names instead of paths
+
+## 0.1.2
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.26
+
+## 0.1.1
+
+### Patch Changes
+
+- b776572: update example repos
+
+## 0.1.0
+
+### Minor Changes
+
+- c3211c6: init versioning
+
+## null
+
+### Patch Changes
+
+- @inlang/settings-component@1.0.25
+
 ## null
 
 ### Patch Changes