+++ inlang/source-code/github-lint-action/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @inlang/github-lint-action
 
+## 0.3.19
+
+### Patch Changes
+
+- 250d76d: Fix missing comment if PR already has a comment
+
+## 0.3.18
+
+### Patch Changes
+
+- 123d206: Add comment logging
+
+## 0.3.17
+
+### Patch Changes
+
+- Updated dependencies [e8d74b9]
+  - @lix-js/fs@2.2.0
+  - @inlang/sdk@0.36.3
+  - @lix-js/client@2.2.1
+
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 0.3.15
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
+## 0.3.13
+
+### Patch Changes
+
+- Updated dependencies [8e9fc0f]
+  - @inlang/sdk@0.35.9
+
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [da7c207]
+  - @inlang/sdk@0.35.8
+
 ## 0.3.11
 
 ### Patch Changes