+++ inlang/source-code/plugins/next-intl/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @inlang/plugin-next-intl
 
+## 1.3.49
+
+### Patch Changes
+
+- c26aac4: refactor icon
+
+## 1.3.48
+
+### Patch Changes
+
+- @inlang/sdk@0.36.3
+
+## 1.3.47
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 1.3.46
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 1.3.45
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
+## 1.3.44
+
+### Patch Changes
+
+- Updated dependencies [8e9fc0f]
+  - @inlang/sdk@0.35.9
+
+## 1.3.43
+
+### Patch Changes
+
+- 04e804b: add human readble id tests to plugins
+- aa91877: add more extract options
+
+## 1.3.42
+
+### Patch Changes
+
+- Updated dependencies [da7c207]
+  - @inlang/sdk@0.35.8
+
+## 1.3.41
+
+### Patch Changes
+
+- 2b36271: add changeset
+
 ## 1.3.40
 
 ### Patch Changes